+++ include/parse_arguments.h
@@ -25,7 +25,7 @@ argparse::ArgumentParser parsed_arguments(
   bool &plot_grid,
   bool &plot_grid_heatmap,
   bool &plot_intersections,
-  bool &crop,
+  bool &crop_polygons,
   bool &plot_polygons,
   bool &remove_tiny_polygons,
   double &minimum_polygon_area,
